+++ Sources/Tabman/AutoInsetter/InsetStore/InsetStore.swift
@@ -3,12 +3,12 @@
 //  AutoInsetter
 //
 //  Created by Merrick Sapsford on 19/03/2019.
-//  Copyright © 2019 UI At Six. All rights reserved.
+//  Copyright © 2022 UI At Six. All rights reserved.
 //
 
 import UIKit
 
-protocol InsetStore: class {
+protocol InsetStore: AnyObject {
     
     func store(contentInset: UIEdgeInsets, for view: UIScrollView)
     func store(contentOffset: CGPoint, for view: UIScrollView)