+++ Sources/Tabman/Bar/BarItem/UIKit+TMBarItemable.swift
@@ -3,7 +3,7 @@
 //  Tabman
 //
 //  Created by Merrick Sapsford on 19/10/2018.
-//  Copyright © 2019 UI At Six. All rights reserved.
+//  Copyright © 2022 UI At Six. All rights reserved.
 //
 
 import UIKit
@@ -14,6 +14,13 @@ extension UINavigationItem: TMBarItemable {
     // swiftlint:disable unused_setter_value
 
     public var image: UIImage? {
+        get {
+            return nil
+        }
+        set {}
+    }
+    
+    public var selectedImage: UIImage? {
         get {
             return nil
         }